refactor: Move packet type validation into `packet` module.
