+++ src/message/outgoing-message.js
@@ -1,26 +1,8 @@
 const Transform = require('readable-stream').Transform;
 
-const { Packet, HEADER_LENGTH, TYPE } = require('../packet');
+const { Packet, HEADER_LENGTH, isValidType } = require('../packet');
 
 const BufferList = require('bl');
-
-function isValidType(type) {
-  switch (type) {
-    case TYPE.SQL_BATCH:
-    case TYPE.RPC_REQUEST:
-    case TYPE.TABULAR_RESULT:
-    case TYPE.ATTENTION:
-    case TYPE.BULK_LOAD:
-    case TYPE.TRANSACTION_MANAGER:
-    case TYPE.LOGIN7:
-    case TYPE.NTLMAUTH_PKT:
-    case TYPE.PRELOGIN:
-      return true;
-
-    default:
-      return false;
-  }
-}
 
 /**
   OutgoingMessage
@@ -32,7 +14,7 @@ module.exports = class OutgoingMessage extends Transform {
   constructor(type, resetConnection, packetSize, debug) {
     super();
 
-    if (typeof type !== 'number' || !isValidType(type)) {
+    if (!isValidType(type)) {
       throw new TypeError('"type" must be a a supported TDS message type');
     }
